# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus API
  version: "1.0.0"
  description: >-
    Argus's read API (docs/SPEC.md §4): agent-telemetry sessions, timelines,
    events, tool calls, subagent trees, and rollup-backed analytics, plus the
    ops endpoints and the ingest surface (OTLP + the hooks webhook). This
    file is hand-authored and reviewed like code (§4.4): it is the contract
    handlers must satisfy, not a byproduct of them.

    Every vendor-supplied field (project, model, vendor, decision_source,
    query_source, tool_source, permission_mode, start_type, terminal_type,
    and friends) is typed `string` with no `enum`, on purpose (SPEC §0,
    §4.4): Argus ingests values it has never seen and must never reject or
    mistype them. Only Argus's own four closed taxonomies — `kind`,
    `source`, `correlation`, `status` (SPEC §0) — and Argus-invented request
    parameters (sort keys, dimensions, buckets) are enumerated.
servers:
  - url: /
    description: Argus serves the ops, read, and ingest surfaces from one origin.
...
paths:
...
  /api/v1/meta:
    get:
      operationId: getMeta
      tags: [meta]
      summary: Version, retention, vendors seen, data-quality summary
      responses:
        "200":
          description: Server metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
              examples:
                default:
                  value:
                    version: "1.4.0"
                    commit: "a1b2c3d"
                    retention_days: 90
                    feature_flags: { estimated_cost: true }
                    vendors: ["claude_code", "codex"]
                    logs_exporter_seen: true
                    metrics_exporter_seen: true
                    hooks_seen: true
                    tool_details_seen: false
                    estimated_cost_present: false
                    data_quality:
                      logs_exporter_seen: true
                      metrics_exporter_seen: true
                      hooks_seen: true
                      tool_details_seen: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/v1/sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List sessions
      parameters:
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/Vendor"
        - $ref: "#/components/parameters/Model"
        - name: status
          in: query
          description: Repeated params OR (SPEC §4.1).
          schema:
            type: array
            items:
              $ref: "#/components/schemas/SessionStatus"
          style: form
          explode: true
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/DecisionSource"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - name: q
          in: query
          description: Substring match on id/project/cwd.
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [last_event_at, started_at, cost_usd, event_count]
            default: last_event_at
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionsListResponse"
              examples:
                default:
                  value:
                    data:
                      - id: "3f7a3b1e-0000-0000-0000-000000000001"
                        vendor: "claude_code"
                        project: "argus"
                        cwd: "/home/y/Labs/argus"
                        status: "active"
                        start_type: "fresh"
                        started_at: "2026-08-11T09:02:11.412Z"
                        ended_at: null
                        last_event_at: "2026-08-11T09:31:44.900Z"
                        duration_ms: 1773488
                        turn_count: 12
                        event_count: 480
                        tool_call_count: 96
                        tool_reject_count: 3
                        subagent_count: 2
                        error_count: 1
                        tokens: { input: 41233, output: 18944, cache_read: 1204331, cache_creation: 88210 }
                        cost:
                          usd: 4.2711
                          reported_usd: 4.2711
                          estimated_usd: 0
                          estimated_share: 0
                          by_query_source: { "": 3.9011, "sdk": 0.35, "generate_session_title": 0.02 }
                          dominant_query_source: ""
                          other_query_source_usd: 0.37
                        models: ["claude-opus-5", "claude-sonnet-4-5"]
                        partial: false
                        app_version: "2.1.220"
                        entrypoint: "cli"
                        terminal_type: "wsl-Ubuntu"
                    page: { next_cursor: "eyJrIjoi4oCm", has_more: true }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/v1/sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Session detail with aggregates
      parameters:
        - $ref: "#/components/parameters/SessionID"
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: The session, plus permission history, top tools, and decision summary.
          headers:
            ETag:
              description: Hash of the underlying max(ts, seq) + filter (SPEC §4.1).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionDetail"
              examples:
                default:
                  value:
                    id: "3f7a3b1e-0000-0000-0000-000000000001"
                    vendor: "claude_code"
                    project: "argus"
                    cwd: "/home/y/Labs/argus"
                    status: "active"
                    start_type: "fresh"
                    started_at: "2026-08-11T09:02:11.412Z"
                    ended_at: null
                    last_event_at: "2026-08-11T09:31:44.900Z"
                    duration_ms: 1773488
                    turn_count: 12
                    event_count: 480
                    tool_call_count: 96
                    tool_reject_count: 3
                    subagent_count: 2
                    error_count: 1
                    tokens: { input: 41233, output: 18944, cache_read: 1204331, cache_creation: 88210 }
                    cost:
                      usd: 4.2711
                      reported_usd: 4.2711
                      estimated_usd: 0
                      estimated_share: 0
                      by_query_source: { "": 3.9011, "sdk": 0.35 }
                      dominant_query_source: ""
                      other_query_source_usd: 0.35
                    models: ["claude-opus-5"]
                    partial: false
                    app_version: "2.1.220"
                    entrypoint: "cli"
                    terminal_type: "wsl-Ubuntu"
                    permission_mode_history:
                      - ts: "2026-08-11T09:05:00.000Z"
                        from: "default"
                        to: "acceptEdits"
                        trigger: "user"
                    top_tools:
                      - { tool_name: "Edit", calls: 40, rejects: 2, p50_ms: 120 }
                    decision_summary:
                      accept: 90
                      reject: 6
                      by_source: { config: 60, hook: 10, user_permanent: 10, user_temporary: 5, user_reject: 4, user_abort: 1 }
                      exact_share: 1.0
                    sources_seen: ["otel_log", "hook"]
                    raw_events_expired: false
                    hook_latency: { p50_ms: 9, p95_ms: 41, by_hook_event: { PostToolUse: 9 } }
                    first_seen_at: "2026-08-11T09:02:10.000Z"
                    user: "yohann"
                    organization_id: "org_1"
        "304":
          description: Not modified — the client's If-None-Match matches the current ETag.
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /api/v1/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: Cross-session event search
      parameters:
        - $ref: "#/components/parameters/Kinds"
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/DecisionSource"
        - $ref: "#/components/parameters/PromptID"
        - $ref: "#/components/parameters/AgentID"
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/Vendor"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/Fields"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of events across sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineListResponse"
              examples:
                default:
                  value:
                    data:
                      - event_ref: "MTc2ODk2NzMyNDIyMTAwMDAwMDo5MTgyMzM"
                        seq: 918233
                        id: "0192abcd-0000-0000-0000-000000000001"
                        ts: "2026-08-11T09:12:04.221Z"
                        session_id: "3f7a3b1e-0000-0000-0000-000000000001"
                        prompt_id: "p_88f1"
                        kind: "tool.decision"
                        event_name: "tool_decision"
                        source: "otel_log"
                        vendor: "claude_code"
                        tool_name: "Edit"
                        tool_use_id: "toolu_01A"
                        decision: "reject"
                        decision_source: "user_reject"
                        tool_source: "builtin"
                        query_source: null
                        model: null
                        tokens: null
                        cost: null
                        duration_ms: null
                        success: null
                        error_type: null
                        agent_id: null
                        agent_type: null
                        permission_mode: "default"
                        file_path: "server/internal/store/postgres/store.go"
                        clock_skewed: false
                    page: { next_cursor: null, has_more: false }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/v1/events/{ref}:
    get:
      operationId: getEvent
      tags: [events]
      summary: One event with full attrs (the detail drawer)
      parameters:
        - name: ref
          in: path
          required: true
          description: Opaque event_ref (base64url of ts+seq, SPEC §4.1). Not a lookup by id.
          schema:
            type: string
      responses:
        "200":
          description: The event, with attrs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
              examples:
                default:
                  value:
                    event_ref: "MTc2ODk2NzMyNDIyMTAwMDAwMDo5MTgyMzM"
                    seq: 918233
                    id: "0192abcd-0000-0000-0000-000000000001"
                    ts: "2026-08-11T09:12:04.221Z"
                    session_id: "3f7a3b1e-0000-0000-0000-000000000001"
                    prompt_id: "p_88f1"
                    kind: "tool.decision"
                    event_name: "tool_decision"
                    source: "otel_log"
                    vendor: "claude_code"
                    tool_name: "Edit"
                    tool_use_id: "toolu_01A"
                    decision: "reject"
                    decision_source: "user_reject"
                    tool_source: "builtin"
                    query_source: null
                    model: null
                    tokens: null
                    cost: null
                    duration_ms: null
                    success: null
                    error_type: null
                    agent_id: null
                    agent_type: null
                    permission_mode: "default"
                    file_path: "server/internal/store/postgres/store.go"
                    clock_skewed: false
                    attrs: { "tool_decision.tool_use_id": "toolu_01A", "tool_decision.decision": "reject" }
        "400":
          description: Malformed event_ref.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                invalidEventRef:
                  value:
                    type: "urn:argus:error:invalid-event-ref"
                    title: "Bad Request"
                    status: 400
                    detail: "event_ref is not valid base64url of ts:seq"
                    instance: "/api/v1/events/not-a-ref"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /api/v1/tool-calls:
    get:
      operationId: listToolCalls
      tags: [tool-calls]
      summary: Cross-session tool-call list — the decision-provenance drill-down
      parameters:
        - $ref: "#/components/parameters/Project"
        - $ref: "#/components/parameters/Tool"
        - $ref: "#/components/parameters/DecisionSource"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of tool calls across sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallsListResponse"
              examples:
                default:
                  $ref: "#/components/examples/ToolCallsListExample"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
components:
  parameters:
    SessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Default 50, max 500 (SPEC §4.1).
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from a previous page's page.next_cursor.
      schema:
        type: string
    From:
      name: from
      in: query
      description: RFC 3339 timestamp or relative shorthand (-24h, -7d). SPEC §4.1.
      schema:
        type: string
    To:
      name: to
      in: query
      description: RFC 3339 timestamp or relative shorthand (-24h, -7d). SPEC §4.1.
      schema:
        type: string
    Project:
      name: project
      in: query
      description: Repeated params OR within the field (SPEC §4.1).
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Model:
      name: model
      in: query
      description: Repeated params OR within the field (SPEC §4.1).
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Vendor:
      name: vendor
      in: query
      description: Repeated params OR within the field (SPEC §4.1).
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Tool:
      name: tool
      in: query
      description: Repeated params OR within the field (SPEC §4.1).
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    DecisionSource:
      name: decision_source
      in: query
      description: Repeated params OR within the field (SPEC §4.1).
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Kinds:
      name: kinds
      in: query
      description: Repeated params OR within the field (SPEC §4.1). Argus's own closed Kind taxonomy.
      schema:
        type: array
        items:
          $ref: "#/components/schemas/Kind"
      style: form
      explode: true
    PromptID:
      name: prompt_id
      in: query
      schema:
        type: string
    AgentID:
      name: agent_id
      in: query
      schema:
        type: string
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    Fields:
      name: fields
      in: query
      description: slim (default, no attrs) or full.
      schema:
        type: string
        enum: [slim, full]
        default: slim
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: ETag from a previous response; a match returns 304 with an empty body.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Malformed or out-of-range request parameter.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            invalidCursor:
              value:
                type: "urn:argus:error:invalid-cursor"
                title: "Bad Request"
                status: 400
                detail: "cursor failed signature check"
                instance: "/api/v1/sessions"
            invalidParameter:
              value:
                type: "urn:argus:error:invalid-parameter"
                title: "Bad Request"
                status: 400
                detail: "from: not a valid RFC 3339 timestamp or relative shorthand: \"garbage\""
                instance: "/api/v1/sessions"
    Unauthorized:
      description: Missing or invalid bearer token (ARGUS_API_TOKEN configured).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            default:
              value:
                type: "urn:argus:error:unauthorized"
                title: "Unauthorized"
                status: 401
                detail: "missing or invalid bearer token"
                instance: "/api/v1/sessions"
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            default:
              value:
                type: "urn:argus:error:not-found"
                title: "Not Found"
                status: 404
                detail: "no such resource"
                instance: "/api/v1/sessions/does-not-exist"
    MethodNotAllowed:
      description: Method not allowed on this resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            default:
              value:
                type: "urn:argus:error:method-not-allowed"
                title: "Method Not Allowed"
                status: 405
                detail: "method not allowed on this resource"
                instance: "/api/v1/sessions"
  examples:
    ToolCallsListExample:
      value:
        data:
          - id: "5b1f7a2e-0000-0000-0000-000000000001"
            session_id: "3f7a3b1e-0000-0000-0000-000000000001"
            prompt_id: "p_88f1"
            tool_use_id: "toolu_01A"
            tool_name: "Edit"
            tool_source: "builtin"
            agent_id: null
            decision: "reject"
            decision_source: "user_reject"
            permission_mode: "default"
            started_at: "2026-08-11T09:12:04.000Z"
            decided_at: "2026-08-11T09:12:04.200Z"
            ended_at: "2026-08-11T09:12:04.200Z"
            duration_ms: 200
            wait_ms: 200
            success: false
            error_type: null
            file_path: "server/internal/store/postgres/store.go"
            input_size_bytes: 412
            result_size_bytes: null
            correlation: "exact"
            event_count: 2
        page: { next_cursor: null, has_more: false }
  schemas:
...
    Problem:
      type: object
      description: >-
        RFC 9457 problem+json (SPEC §4.1). type is a stable
        urn:argus:error:<slug> from Argus's own error catalog (see
        server/internal/httpapi/problem.go); left as a plain string rather
        than enumerated here so a future URN never needs a spec-only
        migration to become valid.
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: true
    PageInfo:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, "null"]
        has_more:
          type: boolean
...
    EventSource:
      type: string
      description: Provenance pipeline (SPEC §1.3) — Argus-closed, one of the four permitted taxonomies.
      enum: [otel_log, otel_metric, hook, sim]
...
    SessionStatus:
      type: string
      description: Argus-computed session lifecycle state (SPEC §1.7) — closed, unlike vendor vocabulary.
      enum: [active, ended, abandoned, unknown]
...
    SessionsListResponse:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        page:
          $ref: "#/components/schemas/PageInfo"
    PermissionModeChange:
      type: object
      required: [ts, from, to, trigger]
      properties:
        ts: { type: string, format: date-time }
        from:
          type: string
          description: Vendor-supplied permission_mode value; unconstrained.
        to:
          type: string
          description: Vendor-supplied permission_mode value; unconstrained.
        trigger:
          type: string
    ToolUsageSummary:
      type: object
      required: [tool_name, calls, rejects, p50_ms]
      properties:
        tool_name: { type: string }
        calls: { type: integer }
        rejects: { type: integer }
        p50_ms:
          type: [integer, "null"]
    SessionDecisionSummary:
      type: object
      required: [accept, reject, by_source, exact_share]
      properties:
        accept: { type: integer }
        reject: { type: integer }
        by_source:
          type: object
          description: Keys are raw decision_source values, unconstrained (SPEC §1.9).
          additionalProperties:
            type: integer
        exact_share: { type: number }
    SessionHookLatency:
      type: object
      required: [p50_ms, p95_ms, by_hook_event]
      properties:
        p50_ms: { type: integer, format: int64 }
        p95_ms: { type: integer, format: int64 }
        by_hook_event:
          type: object
          additionalProperties:
            type: integer
            format: int64
    SessionDetail:
      allOf:
        - $ref: "#/components/schemas/SessionSummary"
        - type: object
          required:
            [permission_mode_history, top_tools, decision_summary, sources_seen,
              raw_events_expired, hook_latency, first_seen_at, user, organization_id]
          properties:
            permission_mode_history:
              type: array
              items:
                $ref: "#/components/schemas/PermissionModeChange"
            top_tools:
              type: array
              items:
                $ref: "#/components/schemas/ToolUsageSummary"
            decision_summary:
              $ref: "#/components/schemas/SessionDecisionSummary"
            sources_seen:
              type: array
              items:
                $ref: "#/components/schemas/EventSource"
            raw_events_expired:
              type: boolean
            hook_latency:
              oneOf:
                - $ref: "#/components/schemas/SessionHookLatency"
                - type: "null"
            first_seen_at:
              type: string
              format: date-time
            user:
              type: string
            organization_id:
              type: string
...
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/TimelineEvent"
        - type: object
          required: [attrs]
          properties:
            attrs:
              type: object
              description: Full flattened source payload, including the promoted fields above (SPEC §1.3).
              additionalProperties: true
    TimelineListResponse:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        page:
          $ref: "#/components/schemas/PageInfo"
...
    ToolCallsListResponse:
      type: object
      required: [data, page]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        page:
          $ref: "#/components/schemas/PageInfo"
...
    DataQuality:
      type: object
      required: [logs_exporter_seen, metrics_exporter_seen, hooks_seen, tool_details_seen]
      properties:
        logs_exporter_seen: { type: boolean }
        metrics_exporter_seen: { type: boolean }
        hooks_seen: { type: boolean }
        tool_details_seen: { type: boolean }
    Meta:
      type: object
      description: >-
        SPEC §4.3 full shape. The live handler (server/internal/httpapi/meta.go)
        currently returns only {version, commit, retention_days} — P3-08
        extends it to this full contract.
      required:
        [version, commit, retention_days, feature_flags, vendors, logs_exporter_seen,
          metrics_exporter_seen, hooks_seen, tool_details_seen, estimated_cost_present, data_quality]
      properties:
        version: { type: string }
        commit: { type: string }
        retention_days: { type: integer }
        feature_flags:
          type: object
          additionalProperties:
            type: boolean
        vendors:
          type: array
          items:
            type: string
        logs_exporter_seen: { type: boolean }
        metrics_exporter_seen: { type: boolean }
        hooks_seen: { type: boolean }
        tool_details_seen: { type: boolean }
        estimated_cost_present: { type: boolean }
        data_quality:
          $ref: "#/components/schemas/DataQuality"
...